refactor(server): peel media routes into routers/media.py + extract mediarecords.py (router-split, round-5 #51 / R5-25 PR25)
Fourteenth router peel. Moves the 16-route media surface into a leaf
routers/media.py: the list (/api/media, filtered/sorted/paginated + semantic/SQL
search branches), position/pool, single-record detail, and the per-clip
sub-resources (waveform/scenes/chapters/rating/tags×3/remotion-props/single-clip
retranscribe/transcripts/transcript-activate/retry-vision), with their 4 models.

The two bulk-fetch helpers shared with the search group — _get_tags_bulk (H15) and
_get_light_records_by_ids (H16) — are extracted to a new leaf mediarecords.py
(db-only), re-exported from server.py so structured_query/search_all (which stay
until the search peel) keep working. Route ORDER is preserved: /api/media/pool and
/api/media/position/{media_id} are declared before the dynamic /api/media/{media_id}
so the typed-int route can't shadow the literal pool segment. config.BASE_DIR
replaces server.ROOT for the waveform cache dir.

Behaviour-preserving: paths/methods/scopes unchanged (SPA + CLI unaffected).
server.py 1857 → 1068. Two direct-reference tests repointed to routers.media
(retranscribe/retry-vision _resolve_media_path patches; mediatypes ext identity).
